extract interfaces to remove circular dependency
