+++ src/common/DataSetSelector.ts
@@ -14,29 +14,8 @@ import {AppContext} from 'phovea_core';
 import {TimePointUtils} from './TimePointUtils';
 import {ProductIDType} from 'phovea_core';
 import {ParseRangeUtils} from 'phovea_core';
-import {Moment} from 'moment';
-import Format = d3.time.Format;
+import {ITacoDataset, ITacoTimePoint} from './interfaces';
 
-export interface ITacoDataset {
-  key: string;
-  values: ITacoTimePoint[];
-}
-
-export interface ITacoTimePoint {
-  item: INumericalMatrix;
-  key: string;
-  time: Moment;
-  timeFormat: ITacoTimeFormat;
-  values: INumericalMatrix[];
-  rowStratId: string;
-  colStratId: string;
-}
-
-export interface ITacoTimeFormat {
-  d3: ((d: any) => string) | Format | string;
-  moment: string;
-  momentIsSame: moment.unitOfTime.StartOf;
-}
 
 class DataProvider {
