+++ src/common/TimePointUtils.ts
@@ -7,7 +7,7 @@ import * as d3 from 'd3';
 import {EventHandler} from 'phovea_core';
 import {AppConstants} from '../app/AppConstants';
 import {AppContext} from 'phovea_core';
-import {ITacoTimePoint} from './DataSetSelector';
+import {ITacoTimePoint} from './interfaces';
 
 export class TimePointUtils {
